Allow user-created widgets
